pyproject.toml
[tool.poetry]
name = "cinema"
version = "0.1.0"
description = ""
authors = ["zhuxiangyualter <1570801561@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
pip = "^24.0"
install = "^1.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
